+++ apps/landing/src/data/testimonials.ts
@@ -3,6 +3,7 @@ interface Testimonial {
   avatar: string;
   name: string;
   review: string;
+  rating: number;
 }
 
 const TESTIMONIALS_DATA: Testimonial[] = [
@@ -10,6 +11,7 @@ const TESTIMONIALS_DATA: Testimonial[] = [
     avatar: 'https://api.dicebear.com/7.x/avataaars/svg?seed=Julie',
     id: 1,
     name: 'Julie Dupont',
+    rating: 5,
     review:
       "Ils ont fait mes courses en moins de temps qu'il ne m'en faut pour choisir un film sur Netflix. Légendaire !",
   },
@@ -17,18 +19,21 @@ const TESTIMONIALS_DATA: Testimonial[] = [
     avatar: 'https://api.dicebear.com/7.x/avataaars/svg?seed=Maxime',
     id: 2,
     name: 'Maxime Leroy',
+    rating: 5,
     review: "Mon chien n'a jamais été aussi heureux. Moi non plus. Merci pour les grasses mat' !",
   },
   {
     avatar: 'https://api.dicebear.com/7.x/avataaars/svg?seed=Marc',
     id: 3,
     name: 'Marc Spector',
+    rating: 5,
     review: "J'ai commandé un ménage express et quand je suis rentré, mon appart brillait plus que mon avenir. Merci !",
   },
   {
     avatar: 'https://api.dicebear.com/7.x/avataaars/svg?seed=Joan',
     id: 4,
     name: 'Arc Joan',
+    rating: 5,
     review: "J'ai enfin rangé mon appart... sans lever le petit doigt. L'application de mes rêves !",
   },
 ];